requirements.txt
streamlit
pillow
pymupdf
easyocr
transformers
torch
g4f